add user model, organize files
