+++ src/test/test-server.js
@@ -1,7 +1,7 @@
 const chai = require('chai');
 const chaiHttp = require('chai-http');
 
-const {app} = require('../server');
+const {app} = require('../../server');
 
 const should = chai.should();
 chai.use(chaiHttp);
